src/socket.c: remove ineffective seteuid() call
The current call to seteuid() has no effect if the effective UID of the
process does not differ from the real UID.

If we wanted a process started as root to drop root privileges we would
need another user to whose UID we could switch after initialization. Such a
user is not foreseen in the current coding.

There is no need to run sispmctl as privileged user. The standard port the
webserver is listening to is above 1024 and the authorization for the USB
device can be granted via udev rules.

So let's remove the ineffective coding.

Signed-off-by: Heinrich Schuchardt <xypron.glpk@gmx.de>